+++ package/src/index.ts
@@ -7,4 +7,4 @@ AudioPlayerWithProviders.CustomComponent = CustomComponent;
 export * from "./components/AudioPlayer";
 export * from "./components/AudioPlayer/Context";
 export * from "./components/AudioPlayer/Player";
-export * from "./components/Provider/SpectrumProvider";
+export * from "./components/Provider/AudioPlayerRootProvider";